Fix clippy complaint, describe pull request in commit message
`IssueAsset` had a bug where it was not validating the asset type in the
`BlindAssetRecord`s against the asset type used for looking up the
issuer key. This meant that any issuer key could be used to issue any
asset.

With this fix, confidential asset types are no longer supported on the
ledger. I think this is an unavoidable implementation limitation until
support for non-confidential -> confidential transfers is added to zei
(this will probably also need to include some shuffling protocol which
provides untraceability).

The Account Simulator and transaction builder have been updated to no
longer generate transactions with confidential asset types.

The Account Simulator has also had some additions:

 - A new simulator, `OneBigTxnAccounts`, has been added. This builds a
   single Transaction containing every every operation. This catches the
   bug patched by `0372c8df4735101aa7ef75dd35eda7fbf3880aaa`.
 - A fixed (and also otherwise improved) `shrink` method for
   AccountScenarios. Previously if a failure was found which triggered with
   either confidential amounts or confidential asset types, shrinking would
   erroneously generate the same AccountsScenario as a shrink option,
   causing the test to infinite loop.
 - A sketch of how to test cross-confidentiality-domain operations, once
   they are supported.