Create routes.md
First draft documentation for our routes.

Eventually, this documentation should be automatically generated one way or another. Currently, we create routes without any support for discoverability.  If we had wrappers for the route creation code, we could add doc comments to each wrapper. When we implement discoverability, we'll want this information available to be displayed in response to incomplete/incorrect URLs.