Fix failure for DefineAsset and IssueAsset operation sent in same txn
The DefineAsset Operation has the raw asset code whereas IssueAsset Operation
has the derived asset code.

Currently if both the operations are sent in the same txn for a new asset it fails
because the Asset code don't match. This is fixed by comparing the derived asset
code from newly defined assets in the txn.